- refactor invoice generation logic to improve handling of subscription statuses and notifications. -  Introduce new traits for managing expired subscriptions and invoice generation
