+++ resources/lang/ar/fms.php
@@ -544,6 +544,13 @@ return [
             'status_change' => 'تغيير الحالة',
             'module_usage' => 'استخدام وحدة',
             'system_event' => 'حدث نظام',
+            'subscription_expired' => 'انتهاء اشتراك',
+            'subscription_near_expiry' => 'اشتراكك يقترب من الانتهاء',
+            'subscription_grace_period' => 'اشتراكك في فترة السماح',
+            'subscription_suspended' => 'تعليق اشتراك',
+            'subscription_reactivated' => 'إعادة تفعيل اشتراك',
+            'subscription_cancelled' => 'إلغاء اشتراك',
+            'subscription_expired' => 'انتهاء اشتراك',
         ],
     ],
     'details' => 'التفاصيل',
